marft/requirements_qwen.txt
# Qwen QLoRA 훈련을 위한 추가 요구사항
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
datasets>=2.14.0
wandb>=0.15.0
tensorboard>=2.14.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
huggingface_hub>=0.16.0
